Refactor creature and game object retrieval in various scripts
- Updated cs_list.cpp to use GetCreaturesBySpawnId and GetGameObjectsBySpawnId for better performance and readability.
- Modified cs_tele.cpp to improve handling of multiple creature data retrieval.
- Enhanced dc_giant_isles_zone.cpp and dc_guildhouse_butler.cpp to utilize new creature and game object retrieval methods.
- Refactored hlbg_reset.cpp to streamline game object reset logic.
- Improved item upgrade scripts to ensure proper type usage and checks.
- Updated mythicplus_run_manager.cpp to use snapshot methods for creature retrieval.
- Enhanced various OutdoorPvP scripts to replace old retrieval methods with more efficient alternatives.
- Added quick full and quick parallel performance test commands in dc_stresstest.cpp for easier testing.

+++ src/server/scripts/DC/ItemUpgrades/ItemUpgradeStatApplication.cpp
@@ -43,7 +43,7 @@ namespace DarkChaos
 
             // Method 5: Trigger item stat update for all equipped items
             // This ensures the item upgrade changes are reflected
-            for (int slot = EQUIPMENT_SLOT_START; slot < EQUIPMENT_SLOT_END; ++slot)
+            for (uint8 slot = EQUIPMENT_SLOT_START; slot < EQUIPMENT_SLOT_END; ++slot)
             {
                 Item* item = player->GetItemByPos(INVENTORY_SLOT_BAG_0, slot);
                 if (item)
